+++ src/tables/nonconst_column.cpp
@@ -5,7 +5,7 @@
 namespace tables
 {
 
-    nonconst_column::nonconst_column(uint32_t rows)
+    nonconst_column::nonconst_column(uint64_t rows)
         : m_default_value(boost::any())
     {
         add_rows(rows);
@@ -21,7 +21,7 @@ namespace tables
         m_default_value = column->default_value();
     }
 
-    boost::any nonconst_column::value(uint32_t row_index) const
+    boost::any nonconst_column::value(uint64_t row_index) const
     {
         assert(row_index < m_values.size());
         return m_values[row_index];
@@ -37,7 +37,7 @@ namespace tables
         return m_default_value;
     }
 
-    void nonconst_column::add_rows(uint32_t rows)
+    void nonconst_column::add_rows(uint64_t rows)
     {
         // Check for overflow
         assert(m_values.size() + rows >= rows);
